@@ -26,63 +34,39 @@
   "author": "Jan-Felix Schwarz",
   "license": "BSD-3-Clause",
   "bugs": {
-    "url": "https://github.com/effektif/react-mentions/issues"
+    "url": "https://github.com/signavio/react-mentions/issues"
   },
-  "homepage": "https://github.com/effektif/react-mentions",
+  "homepage": "https://github.com/signavio/react-mentions",
   "devDependencies": {
-    "babel-cli": "^6.23.0",
-    "babel-core": "^6.22.1",
-    "babel-loader": "^6.2.4",
-    "babel-plugin-transform-class-properties": "^6.11.5",
-    "babel-plugin-transform-runtime": "^6.22.0",
-    "babel-preset-es2015": "^6.22.0",
-    "babel-preset-react": "^6.22.0",
-    "babel-preset-stage-0": "^6.22.0",
-    "bundle-loader": "0.5.1",
-    "chai": "^3.5.0",
-    "chai-enzyme": "^0.6.1",
-    "chai-things": "^0.2.0",
-    "css-loader": "^0.20.0",
-    "envify": "1.2.0",
-    "enzyme": "^2.7.1",
-    "eslint": "^3.16.1",
-    "eslint-config-signavio": "^1.4.2",
-    "eslint-config-signavio-test": "^1.2.1",
-    "grunt": "~1.0.1",
-    "grunt-contrib-copy": "^1.0.0",
-    "grunt-contrib-less": "~1.4.0",
-    "grunt-gh-pages": "~2.0.0",
-    "isparta": "^4.0.0",
-    "jquery": "^2.2.1",
-    "jsdom": "^8.0.4",
-    "less": "~2.3.1",
-    "less-loader": "^2.2.0",
-    "mocha": "^3.2.0",
-    "mversion": "^1.10.0",
-    "radium": "^0.19.1",
-    "react": "^15.3.2",
-    "react-addons-test-utils": "^15.3.2",
-    "react-dom": "^15.3.2",
-    "react-hot-loader": "^3.0.0-beta.6",
-    "react-router": "^2.6.1",
-    "react-test-renderer": "^15.5.4",
-    "recompose": "^0.22.0",
-    "redbox-react": "^1.2.2",
-    "rimraf": "^2.5.2",
-    "sinon": "~1.17.7",
-    "sinon-chai": "^2.8.0",
-    "style-loader": "^0.13.1",
-    "webpack": "2.2.1",
-    "webpack-dev-server": "2.4.1"
+    "babel-eslint": "^7.2.3",
+    "condition-circle": "^2.0.1",
+    "enzyme": "^3.3.0",
+    "enzyme-adapter-react-16": "^1.1.1",
+    "eslint": "^4.1.1",
+    "eslint-config-react-app": "^2.1.0",
+    "eslint-plugin-flowtype": "^2.34.1",
+    "eslint-plugin-import": "^2.6.0",
+    "eslint-plugin-jsx-a11y": "^5.1.1",
+    "eslint-plugin-react": "^7.1.0",
+    "gh-pages": "^1.1.0",
+    "nwb": "0.21.x",
+    "prettier": "^1.10.2",
+    "radium": "^0.21.2",
+    "react": "^16.2.0",
+    "react-dom": "^16.2.0",
+    "recompose": "^0.26.0",
+    "semantic-release": "^12.4.1"
   },
   "dependencies": {
-    "babel-runtime": "^6.22.0",
     "lodash": "^4.5.1",
     "prop-types": "^15.5.8",
     "substyle": "^6.0.0"
   },
   "peerDependencies": {
-    "react": ">=0.14.0",
-    "react-dom": ">=0.14.0"
+    "react": "16.x",
+    "react-dom": "16.x"
+  },
+  "release": {
+    "verifyConditions": "condition-circle"
   }
 }
