@@ -1,3 +1,3 @@
-export default ({
+export default {
   backgroundColor: '#cee4e5',
-})
+}
