@@ -1,16 +1,7 @@
-import React from "react";
-import { expect } from "chai";
-import { mount } from "enzyme";
-
-import { Mention } from "../src";
-import Highlighter from "../src/Highlighter";
-
-describe("Highlighter", () => {
-
-    it("should notify about the current caret position when moutned.");
-    it("should notify about the current care position whenever it changes.");
-    it("should render the current matched mentions.");
-    it("should only show the matched mentions.");
-    it("should be possible to style the mentions.");
-
-});
+describe('Highlighter', () => {
+  it('should notify about the current caret position when mounted.')
+  it('should notify about the current care position whenever it changes.')
+  it('should render the current matched mentions.')
+  it('should only show the matched mentions.')
+  it('should be possible to style the mentions.')
+})
