@@ -1,13 +1,7 @@
-import React from "react";
-import { expect } from "chai";
-import { mount } from "enzyme";
-
-import Suggestion from "../src/Suggestion";
-
-describe("Suggestion", () => {
-
-    it("should be possible to pass a focussed style.");
-    it("should render the current display of the suggestion.");
-    it("should highlight the part of the display that is matched by the current query.");
-
-});
+describe('Suggestion', () => {
+  it('should be possible to pass a focussed style.')
+  it('should render the current display of the suggestion.')
+  it(
+    'should highlight the part of the display that is matched by the current query.'
+  )
+})
