@@ -1,17 +1,9 @@
-import React from "react";
-import { expect } from "chai";
-import { mount } from "enzyme";
-
-import SuggestionsOverlay from "../src/SuggestionsOverlay";
-
-describe("SuggestionsOverlay", () => {
-
-    it("should render a list of all passed suggestions.");
-    it("should be possible to style the list.");
-    it("should be possible to apply styles to the items in the list.");
-    it("should notify when the user clicks on a suggestion.");
-    it("should be possible to show a loading indicator.");
-    it("should be possible to style the loading indicator.");
-    it("should notify when the user enters a suggestion with his mouse.");
-
-});
+describe('SuggestionsOverlay', () => {
+  it('should render a list of all passed suggestions.')
+  it('should be possible to style the list.')
+  it('should be possible to apply styles to the items in the list.')
+  it('should notify when the user clicks on a suggestion.')
+  it('should be possible to show a loading indicator.')
+  it('should be possible to style the loading indicator.')
+  it('should notify when the user enters a suggestion with his mouse.')
+})
